restructure keymap config
